requirements.txt
streamlit
opencv-python
ultralytics
mediapipe
numpy
Pillow
